VSR: Work around availability bug
Bug:

1. Given 3 replica: A0, B1, B2.
2. B2 is partitioned (for the entire scenario).
3. A0 and B1 prepare and commit many messages...
4. A0 commits a checkpoint trigger and checkpoints.
5. B1 crashes before it can commit the trigger or checkpoint.
6. A0 prepares a message.
7. B1 restarts. The very first entry in its WAL is corrupt, but A0 has already overwritten the corresponding entry in its own WAL.

In order to work around the bug, the primary waits from pings from a quorum of backups to determine when it is safe to proceed (i.e. when the backups have checkpointed).
This increases the checkpoint latency spike.
(This won't show up in the benchmark because the benchmark only runs a single replica.)

This hack should be replaced by async checkpoints (or similar) to avoid the latency spike.