[WIP] Scan API Scan, a.k.a. range queries, is a feature that allows one to iterate over a Tree from a range of key_min..key_max (inclusive).
The current implementation is a sketch, capable of iterating over the immutable table and all on-disk levels.
Iteration over the mutable table isn't supported yet.

As entrypoint to test the changes, there is a dirty hack calling scans for debit_account_id on every batch submitted.
This must be replaced with proper unit/fuzz tests, however, any useful use case for scans will be implemented in a separated effort.

Some of the changes that touch Level*Iterator are likely to be reimplemented by other ongoing PRs.

Needs work:

- Tests and fuzz;
- Binary search unit tests;
- Trace and log;